+++ tools/avdecc-get-control.c
@@ -76,7 +76,7 @@ int main( int argc, char **argv )
         arg_descriptor_index = argv[arg];
     }
 
-    sequence_id = strtol( arg_sequence_id, 0, 0 );
+    sequence_id = (uint16_t)strtol( arg_sequence_id, 0, 0 );
     if ( arg_destination_mac )
     {
         jdksavdecc_eui48_init_from_cstr( &destination_mac, arg_destination_mac );
@@ -91,7 +91,7 @@ int main( int argc, char **argv )
         bzero( &target_entity_id, sizeof( target_entity_id ) );
     }
 
-    descriptor_index = strtol( arg_descriptor_index, 0, 0 );
+    descriptor_index = (uint16_t)strtol(arg_descriptor_index, 0, 0);
 
     {
         struct raw_context net;
